[capture] Move cycle converter into scope.py file
As discussed in #243, it makes sense to have the cycle converter
helper functions inside the scope.py file.

Closes #243.
