+++ capture/capture_aes.py
@@ -19,8 +19,8 @@ import yaml
 from Crypto.Cipher import AES
 from lib.ot_communication import OTAES, OTPRNG, OTUART
 from project_library.project import ProjectConfig, SCAProject
-from scopes.cycle_converter import convert_num_cycles, convert_offset_cycles
-from scopes.scope import Scope, ScopeConfig, determine_sampling_rate
+from scopes.scope import (Scope, ScopeConfig, convert_num_cycles,
+                          convert_offset_cycles, determine_sampling_rate)
 from tqdm import tqdm
 
 import util.helpers as helpers
